agregando funcionalidades:             ▪ ahora cada rol tendrán permisos importantes, el rol SuperAdmin  tiene todos los permisos por defecto, el rol Administrador tiene solo dos(2) permisos que son:( crear productos y crear usuarios ), el rol Operario tiene un (1) permiso que es: ( crear noticias ), y el rol Funcionario tiene un (1) permiso que es: ( crear boletines ).
            ▪ Con este nuevo flujo los tres(3) roles que son el Administrador, Operario y el Funcionario tienen permisos por defecto. El objetivo principal es que ayuda a la distribución de ( importar masivos usuarios ).

            ▪ Estos usuarios obtendrán los roles y permisos por defecto, puesto que, evita la selección de roles y permisos uno (1) por uno(1). Por lo tanto, también ayuda a que los nuevos usuarios con roles y permisos, puedan realizarsu trabajo con los permisos que ya tienen por defecto. Y si por algun caso el usuario tiene otras tareas por realizar y no las tenga en su lista de permisos, el rol SuperAdmin o el rol Administrador les dé los permisos. Por lo tanto, en esta nueva configuración, si se pudo obtener los permisos marcados por el rol que pertenece a cada unos. Además tambien se implementó funciones como un usuario con el rol Operario o Funcionario no puede editar, crear, actualizar y desactivar a otro usuario. Tambien una funcion para el usuario con el rol administrador no puede editar, actualizar o desactivar a otro administrador ni un SuperAdmin. Tambien una funcion de NO auto-desactivarse ya sea un Administrador, SuperAdmin y si es el caso, por equivocacion un Operario o un Funcionario.

+++ app/Http/Controllers/BoletinController.php
@@ -1,23 +1,22 @@
 <?php
 
-// actualizacion 09/04/2025 (y ahora con BoletinPolicy 06/06/2025)
-
 namespace App\Http\Controllers;
 
 use App\Models\User;
 use App\Models\Boletin;
-use Illuminate\Support\Str; // Para usar Str::limit en la vista parcial
 use Illuminate\Http\Request;
 use App\Services\BoletinService;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use App\Mail\NuevaRevisionPendienteMail;
 class BoletinController extends Controller
 {
     public function index(Request $request, BoletinService $boletinService)
     {
+        Gate::authorize('crear boletin');
         $boletines = $boletinService->obtenerBoletinFiltrados($request);
         return view('boletines.index', compact('boletines'));
     }
@@ -40,6 +39,7 @@ class BoletinController extends Controller
 
     public function edit(Boletin $boletin)
     {
+        Gate::authorize('editar boletin');
         return view('boletines.edit', compact('boletin'));
     }
 
@@ -63,7 +63,7 @@ class BoletinController extends Controller
             'archivo' => $filePath,
         ]);
 
-        $operadores = User::role('operador')->get();
+        $operadores = User::role('Operario')->get();
         foreach ($operadores as $operador) {
             Mail::to($operador->email)->send(new NuevaRevisionPendienteMail($boletin, 'Boletín'));
         }
@@ -73,6 +73,7 @@ class BoletinController extends Controller
 
     public function update(Request $request, Boletin $boletin)
     {
+        Gate::authorize('actualizar boletin');
         $rules = ([
             'contenido' => 'required|string|max:100',
             'archivo_upload' => 'nullable|file|mimes:pdf|max:5000',
@@ -111,7 +112,7 @@ class BoletinController extends Controller
         $boletin->save();
 
         if ($estadoCambiadoAPendiente) {
-            $operadores = User::role('operador')->get();
+            $operadores = User::role('Operario')->get();
             foreach ($operadores as $operador) {
                 Mail::to($operador->email)->send(new NuevaRevisionPendienteMail($boletin, 'Boletín'));
             }
@@ -134,6 +135,7 @@ class BoletinController extends Controller
 
     public function destroy(Boletin $boletin)
     {
+        Gate::authorize('eliminar boletin');
         if ($boletin->archivo && Storage::disk('public')->exists($boletin->archivo)) {
             Storage::disk('public')->delete($boletin->archivo);
         }
@@ -159,7 +161,7 @@ class BoletinController extends Controller
             'estado' => 'pendiente',
         ]);
 
-        $operadores = User::role('operador')->get();
+        $operadores = User::role('Operario')->get();
         foreach ($operadores as $operador) {
             Mail::to($operador->email)->send(new NuevaRevisionPendienteMail($boletin, 'Boletín'));
         }
