+++ app/Http/Controllers/NoticiaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Noticia; // Importa el modelo Noticia
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Auth; // Para obtener el ID del usuario autenticado
 use Illuminate\Support\Facades\Storage; // Para manejar la carga de imágenes
 use Illuminate\Support\Facades\Response;
@@ -17,6 +18,7 @@ class NoticiaController extends Controller
      */
     public function index(Request $request, NoticiaService $noticiaService)
     {
+        Gate::authorize('crear noticia');
         $noticias = $noticiaService->obtenerNoticiaFiltradas($request);
         // Carga todas las noticias, incluyendo la relación con User para mostrar quién la creó.
         /* $noticias = Noticia::with('user')->get(); */
@@ -77,7 +79,7 @@ class NoticiaController extends Controller
         ]);
 
         // 4. Redirigir al índice de noticias con un mensaje de éxito.
-        return redirect()->route('noticias.noticias.index')->with('success', 'Noticia creada con éxito.');
+        return redirect()->route('noticias.index')->with('success', 'Noticia creada con éxito.');
     }
 
     /**
@@ -97,6 +99,7 @@ class NoticiaController extends Controller
      */
     public function edit(Noticia $noticia)
     {
+        Gate::authorize('editar noticia');
         return view('noticias.edit', compact('noticia'));
     }
 
@@ -106,6 +109,7 @@ class NoticiaController extends Controller
      */
     public function update(Request $request, Noticia $noticia)
     {
+        Gate::authorize('actualizar noticia');
         // 1. Validar los datos del formulario.
         $request->validate([
             'tipo' => 'required|string|max:255',
@@ -142,7 +146,7 @@ class NoticiaController extends Controller
         ]);
 
         // 4. Redirigir al índice de noticias con un mensaje de éxito.
-        return redirect()->route('noticias.noticias.index')->with('success', 'Noticia actualizada con éxito.');
+        return redirect()->route('noticias.index')->with('success', 'Noticia actualizada con éxito.');
     }
 
     /**
@@ -151,6 +155,7 @@ class NoticiaController extends Controller
      */
     public function destroy(Noticia $noticia)
     {
+        Gate::authorize('eliminar noticia');
         // 1. Eliminar la imagen asociada si existe.
         if ($noticia->imagen && Storage::disk('public')->exists($noticia->imagen)) {
             Storage::disk('public')->delete($noticia->imagen);
@@ -160,6 +165,6 @@ class NoticiaController extends Controller
         $noticia->delete();
 
         // 3. Redirigir al índice de noticias con un mensaje de éxito.
-        return redirect()->route('noticias.noticias.index')->with('success', 'Noticia eliminada con éxito.');
+        return redirect()->route('noticias.index')->with('success', 'Noticia eliminada con éxito.');
     }
 }
